node-main/deps/v8/test/webkit/fast/js/Promise-already-resolved-expected.txt

Resolve or reject do not take effect on a resolved Promise.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS result is "foo"
PASS successfullyParsed is true

TEST COMPLETE
